// src/app/interfaces/Transaction.tsx
export enum Category {
  Salary = "Salary",
  Rent = "Rent",
  Groceries = "Groceries",
  Food = "Food",
  InternetPhone = "Internet & Phone",
  HealthInsurance = "Health Insurance",
  Savings = "Savings",
  FixedExp = "Fixed expenses",
  Shopping = "Shopping",
  Entertainment = "Entertainment",
  CarMaintenance = "Car Maintenance",
  KidsSchool = "Kids & School",
  Pets = "Pets",
  GymFitness = "Gym & Fitness",
  StreamingServices = "Streaming Services",
  Home = "Home",
  Investment = "Investment",
  Vacation = "Vacation",
  Birthdays = "Birthdays",
  Christmass = "Christmass",
  Party = "Party",
  Date = "Date",
  Garden = "Garden",
  Other = "Other",
}
// ...
export interface Transaction {
  readonly id: string
  amount: number
  type: "+" | "-"
  date: string
  category: Category
  description?: string
}